feat: init commit

+++ total-repo/server/src/main/java/com/d108/project/domain/forum/post/dto/PostCreateDto.java
@@ -4,8 +4,8 @@ import lombok.Getter;
 
 @Getter
 public class PostCreateDto {
-    private Integer boardId;
-    private Integer memberId;
+    private Long boardId;
+    private Long memberId;
     private String title;
     private String content;
 }
