+++ total-repo/server/src/main/java/com/d108/project/domain/forum/post/repository/PostRepository.java
@@ -1,7 +1,7 @@
 package com.d108.project.domain.forum.post.repository;
 
-import com.d108.project.domain.forum.post.domain.Post;
+import com.d108.project.domain.forum.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
 }
